clean up & refactor the IES loader a bit, split into separate classes, use references instead copies where possible, more consts, turn some static functions into utilities as part of IES profile, refactor & rename its createTexture function + add zAngleDegreeRotation argument allowing for rotating normalized direction vector obtained from octahdronUVToDir utility, update examples_tests submodule
